test/osroom-dev/requirements.txt
aiohttp==3.7.2
amqp==5.0.1
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==3.6.3.0
blinker==1.4
cachelib==0.1.1
celery==5.0.1
celery-once==3.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
crcmod==1.7
cryptography==3.2.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.6
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-WTF==0.14.3
geoip2==4.1.0
gevent==20.9.0
greenlet==0.4.17
has==0.0.4
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.0.0
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.2
Pillow==8.0.1
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psutil==5.7.3
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
PyJWT==1.7.1
pymongo==3.11.0
PyNaCl==1.4.0
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
tld==0.12.2
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.2
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
